Optimized SMA function and Added Battery Symbol
Reworked the SMA function, and a Battery Symbol drawing was pushed to the OLED. The "Battery:" text display was removed from the Display. The Battery Symbol will be on the upper left, alongside the battery health percentage on the right. Next, a conditional charging symbol will be added next to the battery drawing.